trpc: use proxy for rsc
