+++ apps/status-page/src/lib/trpc/server.tsx
@@ -1,83 +1,19 @@
 import "server-only";
-import type { AppRouter } from "@openstatus/api";
-import { HydrationBoundary } from "@tanstack/react-query";
-import { dehydrate } from "@tanstack/react-query";
-import { createTRPCClient } from "@trpc/client";
-import {
-  type TRPCQueryOptions,
-  createTRPCOptionsProxy,
-} from "@trpc/tanstack-react-query";
-import { cookies } from "next/headers";
-import { cache } from "react";
+import { createServerHelpers } from "@openstatus/api/src/rsc";
 
+import { auth } from "../auth";
 import { makeQueryClient } from "./query-client";
-import { endingLink, sentryLoggerLink } from "./shared";
 
-// IMPORTANT: Create a stable getter for the query client that
-//            will return the same client during the same request.
-export const getQueryClient = cache(makeQueryClient);
-
-export const trpc = createTRPCOptionsProxy<AppRouter>({
-  queryClient: getQueryClient,
-  client: createTRPCClient({
-    links: [
-      sentryLoggerLink(),
-      endingLink({
-        headers: {
-          "x-trpc-source": "server",
-        },
-        fetch: async (url, options) => {
-          const cookieStore = await cookies();
-          return fetch(url, {
-            ...options,
-            credentials: "include",
-            headers: {
-              ...options?.headers,
-              cookie: cookieStore.toString(),
-            },
-          });
-        },
-      }),
-    ],
-  }),
+export const {
+  trpc,
+  getQueryClient,
+  HydrateClient,
+  prefetch,
+  batchPrefetch,
+  fetchQueryOrNotFound,
+} = createServerHelpers({
+  // Lazy: `../auth` imports this module, so referencing `auth` eagerly would
+  // hit a circular-init TDZ.
+  makeQueryClient,
+  auth: () => auth(),
 });
-
-export function HydrateClient(props: { children: React.ReactNode }) {
-  const queryClient = getQueryClient();
-
-  return (
-    <HydrationBoundary state={dehydrate(queryClient)}>
-      {props.children}
-    </HydrationBoundary>
-  );
-}
-
-// oxlint-disable-next-line typescript/no-explicit-any -- FIXME: remove any
-export function prefetch<T extends ReturnType<TRPCQueryOptions<any>>>(
-  queryOptions: T,
-) {
-  const queryClient = getQueryClient();
-
-  if (queryOptions.queryKey[1]?.type === "infinite") {
-    // oxlint-disable-next-line typescript/no-explicit-any -- FIXME: remove any
-    void queryClient.prefetchInfiniteQuery(queryOptions as any);
-  } else {
-    void queryClient.prefetchQuery(queryOptions);
-  }
-}
-
-// oxlint-disable-next-line typescript/no-explicit-any -- FIXME: remove any
-export function batchPrefetch<T extends ReturnType<TRPCQueryOptions<any>>>(
-  queryOptionsArray: T[],
-) {
-  const queryClient = getQueryClient();
-
-  for (const queryOptions of queryOptionsArray) {
-    if (queryOptions.queryKey[1]?.type === "infinite") {
-      // oxlint-disable-next-line typescript/no-explicit-any -- FIXME: remove any
-      void queryClient.prefetchInfiniteQuery(queryOptions as any);
-    } else {
-      void queryClient.prefetchQuery(queryOptions);
-    }
-  }
-}
