+++ packages/api/package.json
@@ -16,6 +16,7 @@
     "env": "deno run -A env.ts"
   },
   "dependencies": {
+    "@libsql/client": "catalog:",
     "@openstatus/analytics": "workspace:*",
     "@openstatus/assertions": "workspace:*",
     "@openstatus/db": "workspace:*",
@@ -37,20 +38,25 @@
     "@openstatus/regions": "workspace:*",
     "@openstatus/services": "workspace:*",
     "@openstatus/subscriptions": "workspace:*",
+    "@openstatus/theme-store": "workspace:*",
     "@openstatus/tinybird": "workspace:*",
     "@openstatus/upstash": "workspace:*",
     "@openstatus/utils": "workspace:*",
     "@t3-oss/env-core": "catalog:",
+    "@tanstack/react-query": "catalog:",
     "@trpc/client": "catalog:",
     "@trpc/server": "catalog:",
+    "@trpc/tanstack-react-query": "catalog:",
     "@unkey/api": "catalog:",
     "@vercel/blob": "catalog:",
     "date-fns": "catalog:",
+    "drizzle-orm": "catalog:",
     "isomorphic-dompurify": "catalog:",
     "nanoid": "catalog:",
     "nanoid-dictionary": "catalog:",
     "next": "catalog:",
     "random-word-slugs": "catalog:",
+    "react": "catalog:",
     "stripe": "catalog:",
     "superjson": "catalog:",
     "zod": "catalog:"
